refactor: ModalContents 폴더 구조 변경 (#230)
