+++ src/pages/HomePage/ModalContents/Setting/ModalContentsAlert/ButtonAlert/ButtonAlert.tsx
@@ -13,7 +13,7 @@ const ButtonAlert = ({ children, variant = 'primary', className = '', ...props }
 	return (
 		<button
 			{...props}
-			className={`subhead-semibold-18 w-full rounded-[5px] px-[4.8rem] py-[1rem] text-center text-white ${buttonStyle} ${className}`}
+			className={`w-full rounded-[5px] px-[4.8rem] py-[1rem] text-center text-white subhead-semibold-18 ${buttonStyle} ${className}`}
 		>
 			{children}
 		</button>
