+++ src/pages/HomePage/ModalContents/Setting/ModalContentsAlert/types/index.ts
@@ -1,0 +1,4 @@
+export interface AlertModalProps {
+	handleClose?: () => void;
+	userEmail?: string;
+}
